# services/identity-service/docs/swagger.yml
openapi: 3.0.3
info:
  title: Identity Service API
  description: |
    Authentication and authorization microservice with OAuth2 provider linking.
    
    Issues JWT access & refresh tokens (HMAC-SHA256) for stateless API authentication.
    Supports OAuth2 Authorization Server (RSA-256), OAuth2 Login (Google, GitHub), and
    unified account linking — allowing a single user account with multiple authentication methods.
    
    ## Authentication Flow
    
    1. Register via `POST /auth/register`
    2. Verify email via `POST /auth/verify-email`
    3. Authenticate via `POST /auth/authenticate` to get access + refresh tokens
    4. Send access token in header: `Authorization: Bearer <accessToken>`
    5. When access token expires, refresh via `POST /auth/refresh`
    6. Logout via `POST /auth/logout` to revoke all tokens
    
    ## User Profile & OAuth Providers
    
    - Get profile: `GET /auth/me`
    - List linked providers: `GET /auth/me/providers`
    - Unlink provider: `DELETE /auth/me/providers/{provider}`
    
    ## Forgot Password Flow
    
    1. Request reset email via `POST /auth/forgot-password`
    2. Set new password via `POST /auth/reset-forgotten-password`
    
    ## Token Validation (Service-to-Service)
    
    Other microservices can validate JWT tokens via `POST /auth/validate`
    to verify tokens and retrieve user information without parsing JWT themselves.
    
    ## Token Lifecycle
    
    | Token | TTL | Storage | Revocation |
...
  contact:
...
servers:
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: |
        Validates email verification token and activates user account.
        
        Token expires after 30 minutes and is one-time use only.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
            example:
              token: 550e8400-e29b-41d4-a716-446655440000
      responses:
        "200":
          description: Email verified successfully. Account is now active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResponse'
              example:
                message: Email verified successfully
                verified: true
        "400":
          description: Invalid or expired token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                error: Bad Request
                message: Invalid or expired verification token
                path: /auth/verify-email
                timestamp: "2026-02-27T12:00:00Z"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: |
        Sends new verification email to user.
        
        Rate limited to once per 60 seconds to prevent abuse.
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
            example:
              email: john@example.com
      responses:
        "200":
          description: Verification email sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResponse'
              example:
                message: Verification email sent successfully
                verified: false
        "400":
          description: Rate limit exceeded or account already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                error: Bad Request
                message: Verification email was recently sent. Please wait before requesting another.
                path: /auth/resend-verification
                timestamp: "2026-02-27T12:00:00Z"
        "404":
          description: User or email not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      tags:
        - User Profile
      summary: Get current user profile
      description: |
        Returns detailed information about the currently authenticated user,
        including ID, username, email, verification status, account state, roles, and creation time.
        
        Requires a valid access token in the `Authorization` header.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserDto'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                username: john_doe
                email: john@example.com
                emailVerified: true
                accountState: ACTIVE
                roles:
                  - ROLE_USER
                createdAt: 1740240000000
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me/providers:
    get:
      tags:
        - User Profile
      summary: Get linked OAuth providers
      description: |
        Returns a list of all OAuth providers linked to the current user account.
        
        Includes provider name, associated email, link timestamp, and last login time.
        Useful for displaying "Connected Accounts" in user settings.
        
        Requires a valid access token in the `Authorization` header.
      operationId: getLinkedProviders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of linked OAuth providers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OAuthProviderDto'
              example:
                - id: 550e8400-e29b-41d4-a716-446655440001
                  providerName: GOOGLE
                  providerEmail: john@gmail.com
                  linkedAt: 1740239000000
                  lastLoginAt: 1740240000000
                - id: 550e8400-e29b-41d4-a716-446655440002
                  providerName: GITHUB
                  providerEmail: john_dev
                  linkedAt: 1740238000000
                  lastLoginAt: 1740235000000
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me/providers/{provider}:
    delete:
      tags:
        - User Profile
      summary: Unlink OAuth provider
      description: |
        Removes a specific OAuth provider from the current user's account.
        
        A user must maintain at least one authentication method, so the system
        prevents unlinking the last remaining provider to avoid account lockout.
        
        Requires a valid access token in the `Authorization` header.
      operationId: unlinkProvider
      security:
        - BearerAuth: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: OAuth provider name to unlink (GOOGLE, GITHUB, etc.)
          example: GOOGLE
      responses:
        "204":
          description: OAuth provider unlinked successfully.
        "400":
          description: Cannot unlink the last remaining authentication method.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                error: Bad Request
                message: Cannot unlink your last authentication method
                path: /auth/me/providers/GOOGLE
                timestamp: "2026-02-27T12:00:00Z"
        "401":
          description: Missing or invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Provider is not linked to this account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT **access token** obtained from `POST /auth/authenticate` or `POST /auth/refresh`.
        
        Send in header: `Authorization: Bearer <accessToken>`
        
        The access token contains `"type": "access"` — refresh tokens are rejected by API endpoints.
  schemas:
...
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          format: uuid
          description: Email verification token from email link
          example: 550e8400-e29b-41d4-a716-446655440000
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Email address to send verification to
          example: john@example.com
...
    VerificationResponse:
      type: object
      properties:
        message:
          type: string
          example: Email verified successfully
        verified:
          type: boolean
          example: true
...
    CurrentUserDto:
      type: object
      description: Current authenticated user's profile information.
      required:
        - id
        - username
        - email
        - emailVerified
        - accountState
        - roles
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: User's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: User's login username
          example: john_doe
        email:
          type: string
          format: email
          description: User's email address
          example: john@example.com
        emailVerified:
          type: boolean
          description: Whether the email has been verified
          example: true
        accountState:
          type: string
          description: Current account state
          enum:
            - PENDING_VERIFICATION
            - ACTIVE
            - SUSPENDED
            - LOCKED
            - DELETED
          example: ACTIVE
        roles:
          type: array
          items:
            type: string
          description: Set of role names assigned to the user
          example:
            - ROLE_USER
        createdAt:
          type: integer
          format: int64
          description: Account creation timestamp in milliseconds since epoch
          example: 1740240000000
    OAuthProviderDto:
      type: object
      description: An OAuth provider linked to a user account.
      required:
        - id
        - providerName
        - providerEmail
        - linkedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of this provider link
          example: 550e8400-e29b-41d4-a716-446655440001
        providerName:
          type: string
          description: OAuth provider name (GOOGLE, GITHUB, etc.)
          example: GOOGLE
        providerEmail:
          type: string
          format: email
          description: Email address associated with this provider
          example: john@gmail.com
        linkedAt:
          type: integer
          format: int64
          description: Timestamp when provider was linked (milliseconds since epoch)
          example: 1740239000000
        lastLoginAt:
          type: integer
          format: int64
          nullable: true
          description: Timestamp of last login via this provider (null if never used to login)
          example: 1740240000000
...
    ErrorResponse:
      type: object
      description: Standard error envelope returned on 4xx/5xx responses.
      properties:
        status:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Invalid credentials
        path:
          type: string
          example: /auth/authenticate
        timestamp:
          type: string
          format: date-time
          example: "2026-02-27T12:00:00Z"
security:
  - BearerAuth: []